+++ internal/controller/ingress/httpsedge_controller.go
@@ -43,12 +43,12 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/go-logr/logr"
-	ingressv1alpha1 "github.com/ngrok/kubernetes-ingress-controller/api/ingress/v1alpha1"
-	"github.com/ngrok/kubernetes-ingress-controller/internal/controller/controllers"
-	ierr "github.com/ngrok/kubernetes-ingress-controller/internal/errors"
-	"github.com/ngrok/kubernetes-ingress-controller/internal/ngrokapi"
 	"github.com/ngrok/ngrok-api-go/v5"
 	"github.com/ngrok/ngrok-api-go/v5/backends/tunnel_group"
+	ingressv1alpha1 "github.com/ngrok/ngrok-operator/api/ingress/v1alpha1"
+	"github.com/ngrok/ngrok-operator/internal/controller/controllers"
+	ierr "github.com/ngrok/ngrok-operator/internal/errors"
+	"github.com/ngrok/ngrok-operator/internal/ngrokapi"
 )
 
 type routeModuleComparision string
@@ -223,7 +223,7 @@ func (r *HTTPSEdgeReconciler) reconcileRoutes(ctx context.Context, edge *ingress
 		// It's important to note here that we are intentionally ommiting the `route.Backend` for new routes.
 		//  The success or failure of applying a route's modules is then strongly linked the state of its backend.
 		//  Thus, any route with a backend is considered properly configured.
-		//  See https://github.com/ngrok/kubernetes-ingress-controller/issues/208 for additional context.
+		//  See https://github.com/ngrok/ngrok-operator/issues/208 for additional context.
 		if match == nil {
 			routeLog.Info("Creating new route")
 			req := &ngrok.HTTPSEdgeRouteCreate{