+++ internal/controller/ingress/ippolicy_controllers_test.go
@@ -3,8 +3,8 @@ package controllers
 import (
 	"testing"
 
-	ingressv1alpha1 "github.com/ngrok/kubernetes-ingress-controller/api/ingress/v1alpha1"
 	"github.com/ngrok/ngrok-api-go/v5"
+	ingressv1alpha1 "github.com/ngrok/ngrok-operator/api/ingress/v1alpha1"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/utils/ptr"
 )