+++ internal/store/driver.go
@@ -20,11 +20,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 
-	ingressv1alpha1 "github.com/ngrok/kubernetes-ingress-controller/api/ingress/v1alpha1"
-	ngrokv1alpha1 "github.com/ngrok/kubernetes-ingress-controller/api/ngrok/v1alpha1"
+	ingressv1alpha1 "github.com/ngrok/ngrok-operator/api/ingress/v1alpha1"
+	ngrokv1alpha1 "github.com/ngrok/ngrok-operator/api/ngrok/v1alpha1"
 
-	"github.com/ngrok/kubernetes-ingress-controller/internal/annotations"
-	"github.com/ngrok/kubernetes-ingress-controller/internal/errors"
+	"github.com/ngrok/ngrok-operator/internal/annotations"
+	"github.com/ngrok/ngrok-operator/internal/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -77,7 +77,7 @@ func NewDriver(logger logr.Logger, scheme *runtime.Scheme, controllerName string
 
 // WithMetaData allows you to pass in custom metadata to be added to all resources created by the controller
 func (d *Driver) WithMetaData(customMetadata map[string]string) *Driver {
-	ingressMetadata, err := d.setMetadataOwner("kubernetes-ingress-controller", customMetadata)
+	ingressMetadata, err := d.setMetadataOwner("ngrok-operator", customMetadata)
 	if err != nil {
 		d.log.Error(err, "error marshalling custom metadata", "customMetadata", d.ingressMetadata)
 		return d