+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.13.0
+**Full Changelog**: TBD
+
+### Changed
+- Renamed docker image from `ngrok/kubernetes-ingress-controller` to `ngrok/ngrok-operator`.
+
 ## 0.12.0
 **Full Changelog**: https://github.com/ngrok/kubernetes-ingress-controller/compare/kubernetes-ingress-controller-0.11.0...kubernetes-ingress-controller-0.12.0